fix fk type and json settings

+++ src/main/java/com/example/crud/domain/Book.java
@@ -6,5 +6,5 @@ import lombok.Data;
 public class Book {
   Long id;
   String title;
-  int authorId;
+  Long authorId;
 }
